simulador_ponto_v2/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
db-dtypes
shapely
pyproj
